Moving forward towards functionality.
Current goal is to separate the total intermingling of the REPL with the
rest of the system. This is being achieved through the dreaded Checked
Exception of Java. While a *lot* of code won't compile, it'll be a
straightforward transformation to deal with it now, without random
"swank is nil" errors.

Looking forward, there probably needs to be some refactoring of
SwankInterface.java, which is a pretty epic "god class".